refactor: keep Megatron Triton kernel in flat ops layout
