+++ benchmark/README.md
@@ -191,14 +191,19 @@ python ../benchmarks_visualizer.py \
 ### Megatron fused linear cross-entropy
 
 The Megatron FLCE benchmark checks loss and gradient parity before measuring
-the default cuBLAS/Triton implementation and the opt-in Triton, CuTile, and
-CuTe DSL backends:
+the default Triton implementation and the opt-in CuTile and CuTe DSL backends:
 
 ```bash
 cd benchmark/scripts
 python benchmark_megatron_fused_linear_cross_entropy.py \
     --tp-size 4 --token-counts 512 2048 --vocab-sizes 32000 128256 \
-    --providers megatron-compatible liger liger-triton liger-cutile
+    --providers megatron-compatible liger liger-cutile
+
+# 7B/Llama-3-scale output layer
+python benchmark_megatron_fused_linear_cross_entropy.py \
+    --tp-size 4 --token-counts 16384 --hidden-size 4096 \
+    --vocab-sizes 128256 \
+    --providers megatron-compatible liger liger-cutile
 
 # SM100 with nvidia-cutlass-dsl installed
 python benchmark_megatron_fused_linear_cross_entropy.py \
